+++ src/main/java/com/arxyt/dominionsword/pomkotscompat/client/PomkotsEntitiesClient.java
@@ -2,6 +2,7 @@ package com.arxyt.dominionsword.pomkotscompat.client;
 
 import com.arxyt.dominionsword.pomkotscompat.client.renderer.PilebunkerEffectRenderer;
 import com.arxyt.dominionsword.pomkotscompat.client.renderer.RisingBlockRenderer;
+import com.arxyt.dominionsword.pomkotscompat.client.renderer.GroundCrackEffectRenderer;
 import com.arxyt.dominionsword.pomkotscompat.registry.PomkotsEntities;
 import net.minecraftforge.client.event.EntityRenderersEvent;
 
@@ -13,5 +14,6 @@ public final class PomkotsEntitiesClient {
     public static void registerRenderers(EntityRenderersEvent.RegisterRenderers event) {
         event.registerEntityRenderer(PomkotsEntities.PILEBUNKER_EFFECT.get(), PilebunkerEffectRenderer::new);
         event.registerEntityRenderer(PomkotsEntities.RISING_BLOCK.get(), RisingBlockRenderer::new);
+        event.registerEntityRenderer(PomkotsEntities.GROUND_CRACK.get(), GroundCrackEffectRenderer::new);
     }
 }